Move getMonthly to utils
