+++ src/utils/getMonthlyResultsByOption.js
